requirements.txt
# pytorch
-f https://download.pytorch.org/whl/torch_stable.html

# Build Packages
setuptools~=71.1.0
wheel
twine

# Lint & Coverage
flake8
mypy
coverage

# 3rd party packages
mat73~=0.63
tqdm>=4.32.0
torch>=1.7.1
numpy==1.23.5
mne>=0.21.1
pandas~=2.2.2
py7zr~=0.21.1
joblib~=1.4.2
autograd~=1.6.2
scipy~=1.13.1
pymanopt==0.2.5
requests
requests[socks]
pytest
h5py~=3.11.0
scikit_learn
sphinxcontrib-napoleon
skorch~=1.0.0
pooch~=1.8.2
pylsl~=1.16.2
# pip install psychopy failed on mac[py37, py38], due to wxpython build problem
wxPython==4.1.1; sys_platform == 'darwin' and python_version <= '3.8'
pyglet==1.5.27; sys_platform == 'darwin' and python_version <= '3.8'
psychopy == 2022.1.4
mne-bids

Sphinx
sphinxcontrib-napoleon
sphinxcontrib-apidoc
sphinx-rtd-theme
m2r
m2r2

scikit-learn~=1.5.1
matplotlib~=3.9.1
pyserial~=3.5
pillow~=10.4.0
PyQt5~=5.15.11

metabci~=0.1.2
PyYAML~=6.0.1
onnxruntime~=1.18.1
future~=1.0.0